requirements.txt
# Requests
requests>=2.7.0

# YAML
pyyaml>=3.11


beautifulsoup4
dnspython